+++ src/Jobs/Job.php
@@ -1,8 +1,0 @@
-<?php namespace Orchestra\Extension\Jobs;
-
-use Illuminate\Bus\Queueable;
-
-abstract class Job
-{
-    use Queueable;
-}